Added numbers are now always retrieved as Double. Added Char will be retrieved as String. Optimized some irrelevant stuff.

+++ src/main/kotlin/net/jemzart/jsonkraken/helpers/CharHelper.kt
@@ -6,19 +6,17 @@ internal fun Char.isWhiteSpace(): Boolean {
 
 internal fun Char.isISOControlCharacterOtherThanDelete(): Boolean {
 	val codePoint = this.toInt()
-	// optimized from:
-	// return if (this.toInt() == 0x7F) false else this.isISOControl()
 	return codePoint <= 0x9F && (codePoint > 0x7F || codePoint.ushr(5) == 0)
 }
 
 internal fun Char.isHexadecimal(): Boolean {
-	val codePoint = this.toInt()
+	val codePoint = this.toByte()
 	return codePoint in 48..57  // 0-9
 		|| codePoint in 65..70  // A-F
 		|| codePoint in 97..102 // a-f
 }
 
 internal fun Char.isDecimal(): Boolean {
-	val codePoint = this.toInt()
+	val codePoint = this.toByte()
 	return codePoint in 48..57 // 0-9
 }
